feature: Agregado la funcionalidad y query de buscar por carrera y curso. Añadido el manejo de excepciones de la funcionalidad.

+++ src/main/java/com/ingsoft/tf/api_edurents/service/product/ProductService.java
@@ -5,6 +5,6 @@ import com.ingsoft.tf.api_edurents.dto.ProductDTO;
 import java.util.List;
 
 public interface ProductService {
-    List<ProductDTO> getFilteredProducts(Long carreraId, Long cursoID);
+    List<ProductDTO> getFilteredProducts(Integer carreraId, Integer cursoID);
     List<ProductDTO> getAllProducts();
 }
